+++ app/src/main/java/com/example/maxim/myweather/network/today/Main.java
@@ -13,10 +13,10 @@ public class Main {
     private float temp;
     @SerializedName("pressure")
     @Expose
-    private int pressure;
+    private float pressure;
     @SerializedName("humidity")
     @Expose
-    private int humidity;
+    private float humidity;
     @SerializedName("temp_min")
     @Expose
     private float temp_min;
@@ -38,19 +38,19 @@ public class Main {
         this.temp = temp;
     }
 
-    public int getPressure() {
+    public float getPressure() {
         return pressure;
     }
 
-    public void setPressure(int pressure) {
+    public void setPressure(float pressure) {
         this.pressure = pressure;
     }
 
-    public int getHumidity() {
+    public float getHumidity() {
         return humidity;
     }
 
-    public void setHumidity(int humidity) {
+    public void setHumidity(float humidity) {
         this.humidity = humidity;
     }
